Fix unit tests
- Rename `test/` to `tests/` to match convention

- Remove tests which involve network calls or querying
  a database. These are outside the scope of unit tests.
