--- test/test_add_albums.py
@@ -1,5 +1,0 @@
-from src import add_albums
-
-
-def test_create_db():
-    assert 1 == 1
